refactor(tool_input_repair): extract truncation repair + split tests (dirge-4y4l stage 10a)
Begin decomposing the #2 god-file by ranking:
src/agent/agent_loop/tool_input_repair.rs (2035 LOC). Convert the flat
file into a tool_input_repair/ directory module and split out two
concerns.

- truncation.rs (209 LOC) — JSON truncation repair: TruncationRepairResult,
  repair_truncated_json, ends_with_dangling_key. Pure string-level repair
  (closes unterminated strings / open braces / dangling keys when a model
  hits max_tokens mid-tool-call); no dependency on the JSON Schema or the
  validate/repair machinery. Re-exported via `pub use truncation::*` so
  external paths (crate::agent::agent_loop::tool_input_repair::*) are
  unchanged.
- tests.rs (857 LOC) — the ~853-line #[cfg(test)] mod tests, moved
  verbatim. `use super::*` resolves through the re-exports.
- mod.rs (985 LOC, was 2035) — telemetry types, contract hints, semantic
  repair, validate orchestration, error formatting (clusters deferred to
  later stages); the small inline phase2_tests module stays put.

Public API surface (RepairKind/RepairResult/RepairStats/RepairStatsSnapshot/
repair_truncated_json/with_contract_hint/format_structured_error/
validate_and_repair/is_path_field_name) preserved. Behavior-preserving.
-D warnings clean default + windows-default; 2374 tests green.

Refs dirge-6ra2
